# feat: load native C funcs by pure machine codes