--- a/Makefile
+++ b/Makefile
@@ -23,9 +23,9 @@ CPU_avx			:= amd64
 CPU_avx2		:= amd64
 CPU_sse  		:= amd64
 
-TMPL_avx		:= fastint_amd64_test fastfloat_amd64_test native_amd64_test
-TMPL_avx2		:= fastint_amd64_test fastfloat_amd64_test native_amd64_test
-TMPL_sse 		:= fastint_amd64_test fastfloat_amd64_test native_amd64_test
+TMPL_avx		:= fastint_amd64_test fastfloat_amd64_test native_amd64_test recover_amd64_test
+TMPL_avx2		:= fastint_amd64_test fastfloat_amd64_test native_amd64_test recover_amd64_test
+TMPL_sse 		:= fastint_amd64_test fastfloat_amd64_test native_amd64_test recover_amd64_test
 
 CFLAGS_avx		:= -msse -mno-sse4 -mavx -mpclmul -mno-avx2 -mstack-alignment=0 -DUSE_AVX=1 -DUSE_AVX2=0
 CFLAGS_avx2		:= -msse -mno-sse4 -mavx -mpclmul -mavx2 -mstack-alignment=0 -DUSE_AVX=1 -DUSE_AVX2=1