--- a/loader/funcdata_compat.go
+++ b/loader/funcdata_compat.go
@@ -1,5 +1,5 @@
-//go:build go1.16 && !go1.18
-// +build go1.16,!go1.18
+//go:build !go1.16
+// +build !go1.16
 
 /*
  * Copyright 2021 ByteDance Inc.