… .github/workflows/fuzzing-linux-x64.yml
… .github/workflows/push-check-linux-arm64.yml
… .github/workflows/push-check-linux-x64.yml
… .github/workflows/push-check-qemu.yml
… .github/workflows/push-check-windows.yml
… ast/api_amd64_test.go
… ast/search_test.go
… decode_float_test.go
… decode_test.go
… encode_test.go
… fuzz/fuzz_test.go
… go.sum
… internal/abi/abi_legacy_amd64.go
… internal/abi/abi_regabi_amd64.go
… internal/abi/stubs.go
… internal/loader/funcdata.go
… internal/loader/funcdata_go115.go
… internal/loader/funcdata_go116.go
… internal/loader/funcdata_go118.go
… internal/loader/funcdata_go120.go
… internal/loader/loader.go
… internal/loader/loader_go117_test.go
… internal/loader/loader_test.go
… internal/loader/loader_windows.go
… internal/native/avx/fastfloat_amd64_test.go
… internal/native/avx/fastint_amd64_test.go
… internal/native/avx/native_amd64.go
… internal/native/avx/native_amd64.s
… internal/native/avx/native_amd64_test.go
… internal/native/avx/native_export_amd64.go